requirements.txt
# Telegram Network Nurturing Agent - Dependencies

# Core Framework
python-telegram-bot>=20.7
crewai>=0.28.0
crewai-tools>=0.4.0

# API Integrations
gspread>=5.12.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
openai>=1.6.0
google-generativeai>=0.3.0
google-api-python-client>=2.111.0

# Web Search
serpapi>=0.1.5
google-search-results>=2.4.2

# Data Processing
pandas>=2.1.0
Pillow>=10.1.0
pydub>=0.25.1
regex>=2023.10.3

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
tabulate>=0.9.0

# Audio Processing
aiohttp>=3.9.0
aiofiles>=23.2.1

# Async Support
nest-asyncio>=1.5.8
